feat: 특정 경로에서 푸터 및 하단 nav 숨기기
- `/chat` 경로에서는 푸터, 하단nav 보이지 않도록 조건부 렌더링 추가
